[#12921] test(server): Register secrets-provider audit capture on EventBus
Avoid Class.forName + static list capture that raced empty on CI even when
GET /api/secrets/providers returned 200 through HttpAuditFilter.
